refactor: Entity에서 Dto로 이름 수정으로 인한 코드 변경 #68
