+++ data/src/main/java/com/kdjj/data/di/RepositoryModule.kt
@@ -1,8 +1,8 @@
 package com.kdjj.data.di
 
-import com.kdjj.data.recipe.RecipeRepositoryImpl
-import com.kdjj.data.recipeimage.RecipeImageRepositoryImpl
-import com.kdjj.data.recipetype.RecipeTypeRepositoryImpl
+import com.kdjj.data.repository.RecipeRepositoryImpl
+import com.kdjj.data.repository.RecipeImageRepositoryImpl
+import com.kdjj.data.repository.RecipeTypeRepositoryImpl
 import com.kdjj.data.repository.RecipeListRepositoryImpl
 import com.kdjj.domain.repository.RecipeImageRepository
 import com.kdjj.domain.repository.RecipeListRepository
