+++ local/src/main/java/com/kdjj/local/dao/RecipeTypeDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.kdjj.local.model.RecipeTypeEntity
+import com.kdjj.local.dto.RecipeTypeEntity
 
 @Dao
 internal interface RecipeTypeDao {
