+++ local/src/main/java/com/kdjj/local/dataSource/RecipeImageLocalDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.kdjj.local.dataSource
 
-import com.kdjj.data.recipeimage.RecipeImageLocalDataSource
+import com.kdjj.data.datasource.RecipeImageLocalDataSource
 import com.kdjj.local.FileSaveHelper
 import javax.inject.Inject
