+++ local/src/main/java/com/kdjj/local/dataSource/RecipeListLocalDataSourceImpl.kt
@@ -4,7 +4,7 @@ import com.kdjj.data.common.errorMap
 import com.kdjj.data.datasource.RecipeListLocalDataSource
 import com.kdjj.domain.model.Recipe
 import com.kdjj.local.dao.RecipeListDao
-import com.kdjj.local.model.toDomain
+import com.kdjj.local.dto.toDomain
 import java.lang.Exception
 import javax.inject.Inject
