+++ local/src/main/java/com/kdjj/local/di/LocalModule.kt
@@ -6,7 +6,7 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.sqlite.db.SupportSQLiteDatabase
 import com.kdjj.local.database.RecipeDatabase
-import com.kdjj.local.model.RecipeTypeEntity
+import com.kdjj.local.dto.RecipeTypeEntity
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
