+++ local/src/main/java/com/kdjj/local/dto/RecipeDto.kt
@@ -1,4 +1,4 @@
-package com.kdjj.local.model
+package com.kdjj.local.dto
 
 import androidx.room.Embedded
 import androidx.room.Relation
