+++ remote/src/main/java/com/kdjj/remote/dao/FirebaseStorageDaoImpl.kt
@@ -1,4 +1,4 @@
-package com.kdjj.remote
+package com.kdjj.remote.dao
 
 import android.net.Uri
 import com.google.firebase.storage.StorageReference
