+++ remote/src/main/java/com/kdjj/remote/dao/FirestoreDao.kt
@@ -1,4 +1,4 @@
-package com.kdjj.remote
+package com.kdjj.remote.dao
 
 import com.kdjj.domain.model.RecipeType
